Add tests for /api/v1/facts/noscope casbin auth
